+++ spade-ast/src/lib.rs
@@ -460,6 +460,7 @@ pub struct Register {
     pub pattern: Loc<Pattern>,
     pub clock: Loc<Expression>,
     pub reset: Option<(Loc<Expression>, Loc<Expression>)>,
+    pub initial: Option<Loc<Expression>>,
     pub value: Loc<Expression>,
     pub value_type: Option<Loc<TypeSpec>>,
     pub attributes: AttributeList,